+++ goldsrc/bsp/lump.go
@@ -27,7 +27,7 @@ const ( // DO NOT SORT
 	LumpIndexSize
 )
 
-// hard limits courtesy of SDHLT bspfile.h.
+// hard limits courtesy of SDHLT bspfile.h, -1 means infinite.
 const (
 	MaxMapClipNodes    = 0x7FFF
 	MaxMapEdges        = 256000
@@ -42,9 +42,10 @@ const (
 	MaxMapPlanes       = 0x7FFF
 	MaxMapSurfEdges    = 512000
 	MaxMapTexInfo      = 0x7FFF
-	MaxMapTextures     = 4096
-	MaxMapVertices     = 0xFFFF
-	MaxMapVisibility   = 0x800000
+	// VHLT says 4096 but it must be the _number_ of textures, not the lump size. I went over 100 MB and everything was fine.
+	MaxMapTextures   = -1
+	MaxMapVertices   = 0xFFFF
+	MaxMapVisibility = 0x800000
 )
 
 func (t LumpType) EntrySize() int {